Add showdown as markdown formater \n Add handelbars as template engine \n Add select journal in search workflow
